+++ lib/win_ble.dart
@@ -1,3 +1,4 @@
 export 'package:win_ble/src/win_ble.dart';
+export 'package:win_ble/src/win_ble.dart';
 export 'package:win_ble/src/models/ble_characteristic.dart';
 export 'package:win_ble/src/models/ble_device.dart';