+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.0.5
+
+- Added BleState , to get status of Bluetooth Radio
+
 ## 0.0.4
 
 - Fixed writeWithResponse
-- breaking Change : startScan method is void now , to listen to scan result ,listen to WinBle.scanStream 
+- breaking Change : startScan method is void now , to listen to scan result ,listen to WinBle.scanStream
 
 ## 0.0.2